backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
passlib[bcrypt]
pydantic[email]
python-dotenv
pydantic-settings
python-jose[cryptography]
requests